Create clone n KVM vms
Script for manual clones of n KVM vms from a template vm. You need to have the xml and qcow2 directory of the base vm
The advantage of this script is that you can enter a different destination location for your virtual machines.
You can change the directories or vm name to your need.